+++ modules/services/views/src/main/java/au/com/shiftyjelly/pocketcasts/views/multiselect/MultiSelectToolbar.kt
@@ -40,6 +40,7 @@ class MultiSelectToolbar @JvmOverloads constructor(
         multiSelectHelper: MultiSelectHelper<T>,
         @MenuRes menuRes: Int?,
         activity: FragmentActivity,
+        sourceView: SourceView? = null,
     ) {
         setBackgroundColor(context.getThemeColor(UR.attr.support_01))
         if (menuRes != null) {
@@ -52,14 +53,22 @@ class MultiSelectToolbar @JvmOverloads constructor(
                 menu.clear()
 
                 val maxIcons = multiSelectHelper.maxToolbarIcons
-                it.subList(0, maxIcons).forEachIndexed { _, action ->
+
+                val visibleIcons = it.filter { action ->
+                    if (action is MultiSelectEpisodeAction.RemoveListeningHistory) {
+                        sourceView == SourceView.LISTENING_HISTORY
+                    } else {
+                        action.isVisible
+                    }
+                }.take(maxIcons)
+
+                visibleIcons.forEachIndexed { _, action ->
                     val item = menu.add(Menu.NONE, action.actionId, 0, action.title)
                     item.setIcon(action.iconRes)
                     item.setShowAsAction(MenuItem.SHOW_AS_ACTION_ALWAYS)
-                    item.isVisible = action.isVisible
                 }
 
-                overflowItems = it.subList(maxIcons, it.size)
+                overflowItems = it - visibleIcons
 
                 when (multiSelectHelper) {
                     is MultiSelectBookmarksHelper -> {
@@ -95,7 +104,7 @@ class MultiSelectToolbar @JvmOverloads constructor(
                         AnalyticsEvent.MULTI_SELECT_VIEW_OVERFLOW_MENU_SHOWN,
                         AnalyticsProp.sourceMap(multiSelectHelper.source),
                     )
-                    showOverflowBottomSheet(activity.supportFragmentManager, multiSelectHelper)
+                    showOverflowBottomSheet(activity.supportFragmentManager, multiSelectHelper, sourceView)
                 }
                 true
             } else {
@@ -116,9 +125,10 @@ class MultiSelectToolbar @JvmOverloads constructor(
     private fun showOverflowBottomSheet(
         fragmentManager: FragmentManager?,
         multiSelectHelper: MultiSelectEpisodesHelper,
+        sourceView: SourceView?,
     ) {
         if (fragmentManager == null) return
-        val overflowSheet = MultiSelectBottomSheet.newInstance(overflowItems.map { it.actionId })
+        val overflowSheet = MultiSelectBottomSheet.newInstance(overflowItems.map { it.actionId }, shouldShowRemoveListeningHistory = sourceView == SourceView.LISTENING_HISTORY)
         overflowSheet.multiSelectHelper = multiSelectHelper
         overflowSheet.show(fragmentManager, "multiselectbottomsheet")
     }